renaming indexes to itertools
