+++ lyncs_utils/itertools.py
