+++ test/test_itertools.py
